% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CpdaddPreparation.R
\name{CpdaddPreparation}
\alias{CpdaddPreparation}
\title{Function to compute a compound to adduct matrix}
\usage{
CpdaddPreparation(KeggDB = NULL, Info.Add = NULL, do.Par = TRUE, nClust = 2)
}
\arguments{
\item{KeggDB}{Data frame containing a column named Compound with KEGG identifiers and a
column named exact_mass with the corresponding exact masses. If NULL, the
database available in mWISE will be used.}

\item{Info.Add}{Data frame with adducts and in source fragments information. 
The columns should be:
\itemize{
\item{name} with the name of the adduct or fragment
\item{nmol}{ with the number of molecules (i.e., 2M+H: nmol=2 )}
\item{charge}{ with the charge of the adduct or fragment (i.e., M+3H: charge=3)}
\item{massdiff}{ with the mass difference (i.e., M+H: massdiff=1.007276)}
\item{quasi}{ with a 1 if the adduct should be considered as quasi-molecular and a 0, otherwise (Optional)}
\item{polarity}{ with a character vector indicating the polarity of the adduct or fragment. The options are "positive" 
or "negative"}
}}

\item{do.Par}{TRUE if parallel computing is required. Def: TRUE.}

\item{nClust}{Number of clusters that may be used. Def: Number of clusters - 1.}
}
\value{
Function \code{CpdaddPreparation} returns a data frame with the mass-to-charge ratio of all the KEGG compounds and
adducts or fragments provided in Info.Add.
}
\description{
Function \code{CpdaddPreparation} computes a matrix that allows the fast matching of a peak list 
to KEGG database from a data frame containing adducts and fragments information.
}
