% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fella.R
\docType{data}
\name{Fella}
\alias{Fella}
\title{Fella graph for diffusion prioritization}
\format{
An object of class igraph of length 10.
}
\source{
Generated using FELLA R package.
}
\usage{
data(Fella)
}
\description{
This data provides a KEGG knowledge graph with compounds, 
reactions, enzymes, modules and pathways knowledge.
}
\references{
Picart-Armada, S.; Fernandez-Albert, F.; Vinaixa, M.; Yanes, O.; 
Perera-Lluna, A. 
\emph{An R package to enrich metabolomics data: FELLA}, 2018
}
\keyword{datasets}
