% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_eps.R
\name{fitness.eps}
\alias{fitness.eps}
\title{optimization function}
\usage{
fitness.eps(eps, pca.to.tune, k.tuned, data.prep, IData)
}
\arguments{
\item{eps}{Epsilon parameter test}

\item{pca.to.tune}{Pca used for optimization}

\item{k.tuned}{Optimized k value}

\item{data.prep}{object returned by dataPrep function}

\item{IData}{Intensity correlation data}
}
\value{
results for epsilon optimization
}
\description{
Optimization function for epsilon parameter
}
\keyword{internal}
