% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\name{fitness}
\alias{fitness}
\title{optimization function}
\usage{
fitness(k, pca.to.tune, data.prep, IData)
}
\arguments{
\item{k}{Number of eigenvectors to test}

\item{pca.to.tune}{Pca used for optimization}

\item{data.prep}{object returned by dataPrep function}

\item{IData}{Intensity correlation data}
}
\value{
Results for k optimization
}
\description{
Optimization function for k parameter
}
\keyword{internal}
