% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_Rclass_2levels.R
\docType{data}
\name{graph.Rclass.2levels}
\alias{graph.Rclass.2levels}
\title{RClass (compounds-reactions) graph for diffusion prioritization}
\format{
An object of class igraph of length 10.
}
\source{
Generated from RCLASS database.
}
\usage{
data("graph.Rclass.2levels")
}
\description{
This data provides a KEGG knowledge graph with compounds and
reactions.
}
\keyword{datasets}
