% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_Rclass.R
\docType{data}
\name{graph.Rclass}
\alias{graph.Rclass}
\title{RClass graph for diffusion prioritization}
\format{
An object of class igraph of length 10.
}
\source{
Generated from RCLASS database.
}
\usage{
data(graph.Rclass)
}
\description{
This data provides a KEGG knowledge graph with compounds, 
reactions and rclasses.
}
\keyword{datasets}
