% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeggDB.R
\docType{data}
\name{KeggDB}
\alias{KeggDB}
\title{KEGG database}
\format{
An object of class data frame with 15906 rows and 2 columns.
}
\source{
Extracted from KEGG database.
}
\usage{
data(KeggDB)
}
\description{
This data provides a data frame containing the 
KEGG identifiers
and their exact masses.
}
\keyword{datasets}
