% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_range_format.R, R/matchingStage.R,
%   R/mz_ppm_converter.R
\name{matching-funs}
\alias{matching-funs}
\alias{mass.range.format}
\alias{matchingStage}
\alias{mz.ppm.converter}
\title{Functions to match a peak-intensity table to KEGG database}
\usage{
mass.range.format(
  Peak.List,
  force.mass.range = TRUE,
  mass.range.type = "ppm.mode",
  mz.range,
  ppm = 10
)

matchingStage(
  Peak.List,
  force.mass.range = TRUE,
  mass.range.type = "ppm.mode",
  mz.range = NULL,
  ppm = 10,
  polarity = "positive",
  Add.List = NULL,
  Info.Add = NULL,
  do.Par = TRUE,
  nClust
)

mz.ppm.converter(
  tol = 10,
  mzrange = 0.001,
  mz,
  conversion = c("ppm.to.mz", "mz.to.ppm")
)
}
\arguments{
\item{Peak.List}{Data frame containing the LC-MS features.
Columns should contain:
\itemize{
\item{Peak.Id}{ for a peak identifier}
\item{mz}{ for a mass-to-charge ratio value}
\item{rt}{ for the retention time}
\item{Intensities for each sample}
}}

\item{force.mass.range}{Logical. If TRUE, the m/z range is computed using the user-defined ppm tolerance or mz.range (Def: TRUE).}

\item{mass.range.type}{A character indicating if the m/z range computation is performed using tolerance in ppm ("ppm.mode").
or m/z uncertainty ("mz.range") (Def: "ppm.mode").}

\item{mz.range}{If mass.range.type = "mz.range", it indicates the m/z range uncertainty (i.e. mz.range = 0.001).}

\item{ppm}{If mass.range.type = "ppm.mode", it indicates the tolerance in ppm (i.e. ppm = 10).}

\item{polarity}{Acquisition mode of the study. It can be "positive" or "negative".}

\item{Add.List}{List of adducts or fragments to consider.}

\item{Info.Add}{Data frame with adducts information. If NULL, the default information is used.}

\item{do.Par}{Logical. If parallel computing is required (Def: TRUE).}

\item{nClust}{Number of clusters that may be used. Def: Number of clusters - 1.}

\item{tol}{Tolerance in ppm. Def: 10}

\item{mzrange}{mz uncertainty. Def: 0.001}

\item{mz}{mass-to-charge ratio for which the conversion is performed.}

\item{conversion}{Direction of the conversion. Character vector containing some of: "ppm.to.mz", "mz.to.ppm"}
}
\value{
Function \code{mass.range.format} returns the input peak list with two additional columns containing the mzmin and mzmax values.

Function \code{matchingStage} returns a list containing the input peak list and the table of annotated peaks,
containing all the potential KEGG candidates for each LC-MS feature.

Function \code{mz.ppm.converter} returns a vector of values containing uncertainty (if conversion="ppm.to.mz")
or tolerance in ppm (if conversion="mz.to.ppm")
}
\description{
Function \code{mass.range.format} builds mz ranges for a peak-intensity table using \code{mz.ppm.converter} if required.

Function \code{matchingStage} performs a fast matching of the peak-intensity table to KEGG database
using adducts and fragments knowledge.

Function \code{mz.ppm.converter} transforms the user-defined tolerance from ppm to mz uncertainty or viceversa.
}
\examples{
Annotated.Table <- matchingStage(Peak.List, force.mass.range = TRUE, mass.range.type = "ppm.mode",
mz.range = NULL, ppm = 10, polarity, Add.List = NULL, Info.Add = NULL,
do.Par = TRUE, nClust)
}
