% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mWISE_annotation.R
\name{mWISE.annotation}
\alias{mWISE.annotation}
\title{Function to perform the complete annotation pipeline of mWISE}
\usage{
mWISE.annotation(
  Peak.List,
  ppm = 10,
  polarity,
  Info.Add = NULL,
  Add.List = NULL,
  background = NULL,
  diffusion.input.type = "probability",
  Unique.Ann = FALSE,
  Add.Id = c("M+H", "M-H"),
  Rt.05 = 5,
  graph = NULL,
  K = NULL,
  score = "z",
  Intensity.idx,
  graph.name,
  nClust,
  do.Par = TRUE
)
}
\arguments{
\item{Peak.List}{Data frame containing the LC-MS features.
Columns should contain:
\itemize{
\item{Peak.Id}{ for a peak identifier}
\item{mz}{ for a mass-to-charge ratio value}
\item{rt}{ for the retention time}
\item{Intensities for each sample}
}}

\item{ppm}{Tolerance in ppm (Def: 10 ppm).}

\item{polarity}{Acquisition mode of the dataset. It can be "positive" or "negative".}

\item{Info.Add}{A data frame with adducts/fragments knowledge. Columns should be:
\itemize{
\item{name}{ for the adduct or fragment}
\item{mz}{ for a mass-to-charge ratio value}
\item{rt}{ for the retention time}
\item{Intensities for each sample}
}}

\item{nClust}{Number of clusters that may be used. Def: Number of clusters - 1.}

\item{do.Par}{TRUE if parallel computing is required. Def: TRUE}
}
\value{
Function \code{matchingStage} returns a list containing the input peak list and the table of annotated peaks,
containing all the potential KEGG candidates for each LC-MS feature.
}
\description{
Wrapper function that performs the complete workflow of mWISE to annotate a peak-intensity matrix.
}
