% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performanceEvaluation.R
\name{performanceEvaluation}
\alias{performanceEvaluation}
\title{Function to evaluate mWISE performance}
\usage{
performanceEvaluation(
  Annotated.dataset,
  df.Ref,
  top.cmps,
  do.Par = TRUE,
  nClust
)
}
\arguments{
\item{Annotated.dataset}{Object returned by \code{mWISE.annotation} function.}

\item{df.Ref}{Reference data frame. It should contain a column named "Peak.Id"
with a peak identifier that corresponds to the mWISE Peak identifier, and a column named "Kegg"
with the reference KEGG identifiers.}

\item{top.cmps}{Number of top compounds to be considered in performance evaluation.}

\item{do.Par}{TRUE if parallel computing is required. Def: TRUE.}

\item{nClust}{Number of clusters that may be used. Def: Number of clusters - 1.}
}
\value{
Function \code{performanceEvaluation} returns a data frame with the performance of each mWISE stage.
}
\description{
Function \code{performanceEvaluation} computes performance metrics when a reference table is available.
}
