% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_dataset.R
\docType{data}
\name{sample.dataset}
\alias{sample.dataset}
\title{Sample dataset}
\format{
An object of class list with length 2.
}
\source{
Generated using \code{CAMERA} R package from 
the xcms objects extracted 
from ProbMetab website.
}
\usage{
data(sample.dataset)
}
\description{
This data provides a list containing 
information of an LC-MS
data set of Trypanosoma Brucei. It contains an input 
peak list and a reference
table for both positive and negative acquisition modes. 
This is a reduced version
of the whole data set.
}
\keyword{datasets}
