% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_graph.R
\docType{data}
\name{sample.graph}
\alias{sample.graph}
\title{Example graph from FELLA}
\format{
An object of class igraph of length 10.
}
\source{
Extracted from FELLA R package.
}
\usage{
data(sample.graph)
}
\description{
This data provides a sample KEGG knowledge graph with compounds, 
reactions, enzymes, modules and pathways knowledge.
}
\references{
Picart-Armada, S.; Fernandez-Albert, F.; Vinaixa, M.; Yanes, O.; Perera-Lluna, A. 
\emph{An R package to enrich metabolomics data: FELLA}, 2018
}
\keyword{datasets}
