% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_keggDB.R
\docType{data}
\name{sample.keggDB}
\alias{sample.keggDB}
\title{sample KEGG database}
\format{
An object of class data frame with 1017 rows and 2 columns.
}
\source{
Subset of KeggDB.
}
\usage{
data(sample.keggDB)
}
\description{
This data provides a subset of KEGG identifiers
and their exact masses.
}
\keyword{datasets}
