% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trypanosoma_dataset.R
\docType{data}
\name{Trypanosoma.dataset}
\alias{Trypanosoma.dataset}
\title{Example dataset}
\format{
An object of class list with length 2.
}
\source{
Generated using \code{CAMERA} R package from the xcms objects extracted 
from ProbMetab website.
}
\usage{
data(Trypanosoma.dataset)
}
\description{
This data provides a list containing information of an LC-MS
data set of Trypanosoma Brucei. It contains an input peak list and a reference
table for both positive and negative acquisition modes.
}
\keyword{datasets}
